validation for all category service

+++ app/category.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"regexp"
 
 	"github.com/go-playground/validator/v10"
 	"github.com/gorilla/mux"
@@ -40,10 +41,11 @@ func CreateCategory(w http.ResponseWriter, r *http.Request) {
 		fmt.Print(err)
 		w.WriteHeader(400)
 	}
-	errv := validate.Struct(cat)
+	errv := validate.Struct(cat) //create struct validation
 	if errv != nil {
 		fmt.Println(errv)
 		w.WriteHeader(401)
+		json.NewEncoder(w).Encode("Validation error")
 	} else {
 		count, errc := categoryCollection.CountDocuments(context.TODO(), bson.D{{"cid", cat.CId}})
 		fmt.Println("check count of cid:", count)
@@ -76,24 +78,29 @@ func GetCategory(w http.ResponseWriter, r *http.Request) {
 
 	w.Header().Set("Content-Type", "application/json")
 	params := mux.Vars(r)["id"] //get Parameter value as string
-
+	is_alphanumeric := regexp.MustCompile(`^[a-zA-Z0-9]*$`).MatchString(params)
 	// var body category
 	// e := json.NewDecoder(r.Body).Decode(&body)
 	// if e != nil {
 	// 	fmt.Print(e)
 	// }
-	var result primitive.M //  an unordered representation of a BSON document which is a Map
-	filter := bson.M{"cid": params}
-	err := categoryCollection.FindOne(context.TODO(), filter).Decode(&result)
-	if err != nil {
+	if is_alphanumeric {
+		var result primitive.M //  an unordered representation of a BSON document which is a Map
+		filter := bson.M{"cid": params}
+		err := categoryCollection.FindOne(context.TODO(), filter).Decode(&result)
+		if err != nil {
 
-		fmt.Println(err)
-		w.WriteHeader(204)
-		json.NewEncoder(w).Encode("No data found!")
+			fmt.Println(err)
+			w.WriteHeader(204)
+			json.NewEncoder(w).Encode("No data found!")
 
+		} else {
+			json.NewEncoder(w).Encode(result) // returns a Map containing document
+			w.WriteHeader(200)
+		}
 	} else {
-		json.NewEncoder(w).Encode(result) // returns a Map containing document
-		w.WriteHeader(200)
+		json.NewEncoder(w).Encode("Please enter the correct Category ID should be alphanumeric!")
+		w.WriteHeader(204)
 	}
 
 }
@@ -132,12 +139,13 @@ func GetAllCategory(w http.ResponseWriter, r *http.Request) {
 func UpdateCategory(w http.ResponseWriter, r *http.Request) {
 
 	w.Header().Set("Content-Type", "application/json")
+	validate := validator.New()
 
 	type updateBody struct {
-		CId         string `json:"cid"`         //value that has to be matched
-		Cname       string `json:"cname"`       // value that has to be modified
-		Cdesc       string `json:"cdesc"`       // value that has to be modified
-		Cmodifiedby string `json:"cmodifiedby"` // value that has to be modified
+		CId         string `json:"cid" validate:"required,alphanum,min=4,max=10"` //value that has to be matched
+		Cname       string `json:"cname" validate:"required,min=3,max=20"`        // value that has to be modified
+		Cdesc       string `json:"cdesc" validate:"required,min=5,max=100"`       // value that has to be modified
+		Cmodifiedby string `json:"cmodifiedby" validate:"required,min=3,max=20"`  // value that has to be modified
 	}
 	var body updateBody
 	e := json.NewDecoder(r.Body).Decode(&body)
@@ -146,20 +154,29 @@ func UpdateCategory(w http.ResponseWriter, r *http.Request) {
 		fmt.Print(e)
 		w.WriteHeader(400)
 	}
-	filter := bson.D{{"cid", body.CId}} // converting value to BSON type
-	after := options.After              // for returning updated document
-	returnOpt := options.FindOneAndUpdateOptions{
 
-		ReturnDocument: &after,
+	errv := validate.Struct(body) // update struct validation
+
+	if errv != nil {
+		fmt.Println(errv)
+		w.WriteHeader(401)
+		json.NewEncoder(w).Encode("Validation error")
+	} else {
+		filter := bson.D{{"cid", body.CId}} // converting value to BSON type
+		after := options.After              // for returning updated document
+		returnOpt := options.FindOneAndUpdateOptions{
+
+			ReturnDocument: &after,
+		}
+		update := bson.D{{"$set", bson.D{{"cname", body.Cname}, {"cdesc", body.Cdesc}, {"cmodifiedby", body.Cmodifiedby}}}}
+		updateResult := categoryCollection.FindOneAndUpdate(context.TODO(), filter, update, &returnOpt)
+
+		var result primitive.M
+		_ = updateResult.Decode(&result)
+
+		json.NewEncoder(w).Encode(result)
+		w.WriteHeader(200)
 	}
-	update := bson.D{{"$set", bson.D{{"cname", body.Cname}, {"cdesc", body.Cdesc}, {"cmodifiedby", body.Cmodifiedby}}}}
-	updateResult := categoryCollection.FindOneAndUpdate(context.TODO(), filter, update, &returnOpt)
-
-	var result primitive.M
-	_ = updateResult.Decode(&result)
-
-	json.NewEncoder(w).Encode(result)
-	w.WriteHeader(200)
 
 }
 
@@ -168,10 +185,11 @@ func UpdateCategory(w http.ResponseWriter, r *http.Request) {
 func UpdateCategoryStatus(w http.ResponseWriter, r *http.Request) {
 
 	w.Header().Set("Content-Type", "application/json")
+	validate := validator.New()
 
 	type updateBody struct {
-		CId     string `json:"cid"`     //value that has to be matched
-		Cstatus bool   `json:"cstatus"` // value that has to be modified
+		CId     string `json:"cid" validate:"required,alphanum,min=4,max=10"` //value that has to be matched
+		Cstatus bool   `json:"cstatus"`                                       // value that has to be modified
 	}
 	var body updateBody
 	e := json.NewDecoder(r.Body).Decode(&body)
@@ -180,20 +198,28 @@ func UpdateCategoryStatus(w http.ResponseWriter, r *http.Request) {
 		fmt.Print(e)
 		w.WriteHeader(400)
 	}
-	filter := bson.D{{"cid", body.CId}} // converting value to BSON type
-	after := options.After              // for returning updated document
-	returnOpt := options.FindOneAndUpdateOptions{
+	errv := validate.Struct(body) // update status struct validation
 
-		ReturnDocument: &after,
+	if errv != nil {
+		fmt.Println(errv)
+		w.WriteHeader(401)
+		json.NewEncoder(w).Encode("Validation error")
+	} else {
+		filter := bson.D{{"cid", body.CId}} // converting value to BSON type
+		after := options.After              // for returning updated document
+		returnOpt := options.FindOneAndUpdateOptions{
+
+			ReturnDocument: &after,
+		}
+		update := bson.D{{"$set", bson.D{{"cstatus", body.Cstatus}}}}
+		updateResult := categoryCollection.FindOneAndUpdate(context.TODO(), filter, update, &returnOpt)
+
+		var result primitive.M
+		_ = updateResult.Decode(&result)
+
+		json.NewEncoder(w).Encode(result)
+		w.WriteHeader(200)
 	}
-	update := bson.D{{"$set", bson.D{{"cstatus", body.Cstatus}}}}
-	updateResult := categoryCollection.FindOneAndUpdate(context.TODO(), filter, update, &returnOpt)
-
-	var result primitive.M
-	_ = updateResult.Decode(&result)
-
-	json.NewEncoder(w).Encode(result)
-	w.WriteHeader(200)
 
 }
 
@@ -203,19 +229,26 @@ func DeleteCategory(w http.ResponseWriter, r *http.Request) {
 
 	w.Header().Set("Content-Type", "application/json")
 	params := mux.Vars(r)["id"] //get Parameter value as string
+	is_alphanumeric := regexp.MustCompile(`^[a-zA-Z0-9]*$`).MatchString(params)
+
 	// _id, err := primitive.ObjectIDFromHex(params) // convert params to mongodb Hex ID
 	// if err != nil {
 	// 	fmt.Printf(err.Error())
 	// }
-	opts := options.Delete().SetCollation(&options.Collation{}) // to specify language-specific rules for string comparison, such as rules for lettercase
-	res, err := categoryCollection.DeleteOne(context.TODO(), bson.D{{"cid", params}}, opts)
-	if err != nil {
-		log.Fatal(err)
-		w.WriteHeader(400)
+	if is_alphanumeric {
+		opts := options.Delete().SetCollation(&options.Collation{}) // to specify language-specific rules for string comparison, such as rules for lettercase
+		res, err := categoryCollection.DeleteOne(context.TODO(), bson.D{{"cid", params}}, opts)
+		if err != nil {
+			log.Fatal(err)
+			w.WriteHeader(400)
+		} else {
+			fmt.Printf("deleted %v documents\n", res.DeletedCount)
+			json.NewEncoder(w).Encode(res.DeletedCount) // return number of documents deleted
+			w.WriteHeader(200)
+		}
 	} else {
-		fmt.Printf("deleted %v documents\n", res.DeletedCount)
-		json.NewEncoder(w).Encode(res.DeletedCount) // return number of documents deleted
-		w.WriteHeader(200)
+		json.NewEncoder(w).Encode("Please enter the correct Category ID should be alphanumeric!")
+		w.WriteHeader(204)
 	}
 
 }
